third_party/WebKit/LayoutTests/media/autoplay/media-document-autoplay-enabled.html

<!DOCTYPE html>
<title>Tests that a media document can open a media URL with an ambiguous extension.</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../media-file.js"></script>
<body>
<script>
async_test(function(t) {
  const iframe = document.createElement('iframe');

  iframe.onload = t.step_func_done(() => {
    const video = iframe.contentDocument.querySelector('video');
    assert_not_equals(undefined, video);
    assert_false(video.paused);
  });

  iframe.src = findMediaFile('video', '../content/test');
  document.body.appendChild(iframe);
});
</script>
